add review authorization

+++ src/main/java/com/deltadc/quizletclone/review/ReviewService.java
@@ -1,15 +1,19 @@
 package com.deltadc.quizletclone.review;
 
 import com.deltadc.quizletclone.set.SetRepository;
+import com.deltadc.quizletclone.user.User;
 import com.deltadc.quizletclone.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -17,6 +21,16 @@ public class ReviewService {
     private final SetRepository setRepository;
     private final ReviewRepository reviewRepository;
     private final UserRepository userRepository;
+
+    private boolean isReviewOwner(Review r) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = userDetails.getUsername();
+
+        User user = userRepository.findByEmail(username).orElseThrow();
+        Long userId = user.getUser_id();
+
+        return Objects.equals(userId, r.getUser_id());
+    }
 
     public ResponseEntity<Double> getSetAverageRating(@PathVariable Long setId) {
         List<Review> setReviews = reviewRepository.findReviewsBySetId(setId);
@@ -73,6 +87,10 @@ public class ReviewService {
     public ResponseEntity<?> editReviewById(Long reviewId, Review newReview) {
         Review review = reviewRepository.findById(reviewId).orElseThrow();
 
+        if(!isReviewOwner(review)) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Không được sửa review nếu không phải chủ sở hữu");
+        }
+
         review.setTotalStars(newReview.getTotalStars());
 
         reviewRepository.save(review);
@@ -83,9 +101,13 @@ public class ReviewService {
     public ResponseEntity<?> deleteReviewById(Long reviewId) {
         Review review = reviewRepository.findById(reviewId).orElseThrow();
 
+        if(!isReviewOwner(review)) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Không được xóa review nếu không phải chủ sở hữu");
+        }
+
         reviewRepository.delete(review);
 
-        return ResponseEntity.ok("da xoa review");
+        return ResponseEntity.ok("Đã xóa review");
     }
 
     public ResponseEntity<?> getReviewByUserIdAndSetId(Long setId, Long userId) {
